pyproject.toml
[tool.poetry]
name = "pythoneda-infrastructure-git-repositories"
version = "0.0.1a7"
description = "Infrastructure layer of PythonEDA Git Repositories"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-infrastructure/git-repositories"
packages = [{include = "PythonEDAInfrastructureGitRepositories"}]

[tool.poetry.dependencies]
python = "^3.6"
pythoneda = "^0.0.1a7"
pythoneda-git-repositories = "^0.0.1a4"
pythoneda-infrastructure-base = "^0.0.1a5"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
